+++ src/lib/jp2/plugin/plugin_bridge.h
@@ -36,7 +36,7 @@ namespace grk {
 void decode_synch_plugin_with_host(TileProcessor *tcd);
 
 void encode_synch_with_plugin(TileProcessor *tcd, uint32_t compno, uint32_t resno,
-		uint32_t bandno, uint32_t precno, uint32_t cblkno, grk_tcd_band *band,
+		uint32_t bandno, uint64_t precno, uint64_t cblkno, grk_tcd_band *band,
 		grk_tcd_cblk_enc *cblk, uint32_t *numPix);
 
 bool tile_equals(grk_plugin_tile *plugin_tile, grk_tcd_tile *p_tile);
